+++ projects/Mallard/src/components/front/card-group/card.tsx
@@ -1,13 +1,13 @@
-import React, { useMemo } from 'react'
+import React from 'react'
 import { View, StyleSheet, StyleProp, ViewStyle, Image } from 'react-native'
 import { metrics } from '../../../theme/spacing'
 import { withNavigation, NavigationInjectedProps } from 'react-navigation'
 import { Highlight } from '../../highlight'
-import { HeadlineCardText, HeadlineKickerText } from '../../styled-text'
 
 import { useArticleAppearance } from '../../../theme/appearance'
 import { FrontArticle } from '../../../common'
-import { color } from 'src/theme/color'
+
+import { TextBlock } from './text-block'
 
 const styles = StyleSheet.create({
     root: {
@@ -27,85 +27,6 @@ interface PropTypes {
     path: FrontArticle['path']
 }
 
-type TextBlockAppearance = 'default' | 'highlight' | 'pillarColor'
-
-const textBlockStyles = {
-    root: {
-        paddingBottom: metrics.vertical,
-        paddingTop: metrics.vertical / 3,
-    },
-    rootWithHighlight: {
-        backgroundColor: color.palette.highlight.main,
-        paddingHorizontal: metrics.horizontal / 2,
-    },
-    contrastText: {
-        color: color.palette.neutral[100],
-    },
-    headline: {
-        color: color.dimText,
-    },
-}
-
-const useTextBlockStyles = (textBlockAppearance: TextBlockAppearance) => {
-    const { appearance } = useArticleAppearance()
-    switch (textBlockAppearance) {
-        case 'highlight':
-            return {
-                rootStyle: [
-                    textBlockStyles.root,
-                    textBlockStyles.rootWithHighlight,
-                ],
-                kickerStyle: null,
-                headlineStyle: textBlockStyles.headline,
-            }
-        case 'pillarColor':
-            return {
-                rootStyle: [
-                    textBlockStyles.root,
-                    textBlockStyles.rootWithHighlight,
-                    appearance.contrastCardBackgrounds,
-                ],
-                kickerStyle: textBlockStyles.contrastText,
-                headlineStyle: textBlockStyles.contrastText,
-            }
-        default:
-            return {
-                rootStyle: textBlockStyles.root,
-                kickerStyle: [appearance.text, appearance.kicker],
-                headlineStyle: [
-                    textBlockStyles.headline,
-                    appearance.text,
-                    appearance.kicker,
-                ],
-            }
-    }
-}
-
-const TextBlock = ({
-    kicker,
-    headline,
-    appearance,
-}: {
-    kicker: string
-    headline: string
-    appearance: TextBlockAppearance
-}) => {
-    const { rootStyle, kickerStyle, headlineStyle } = useTextBlockStyles(
-        appearance,
-    )
-
-    return (
-        <View style={rootStyle}>
-            <HeadlineKickerText style={kickerStyle}>
-                Kick {kicker}
-            </HeadlineKickerText>
-            <HeadlineCardText style={headlineStyle}>
-                headline {headline}
-            </HeadlineCardText>
-        </View>
-    )
-}
-
 const SmallCard = withNavigation(
     ({
         style,
@@ -114,7 +35,6 @@ const SmallCard = withNavigation(
         navigation,
     }: PropTypes & NavigationInjectedProps<{}>) => {
         const { appearance } = useArticleAppearance()
-        const blockAppearance = 'highlight'
         return (
             <View style={style}>
                 <Highlight
@@ -145,7 +65,7 @@ const SmallCard = withNavigation(
                         <TextBlock
                             kicker={article.kicker}
                             headline={article.headline}
-                            appearance={'pillarColor'}
+                            textBlockAppearance={'pillarColor'}
                         />
                     </View>
                 </Highlight>
